test: remove describe_topic from redpanda service
Remove the sugar interface from the low-level redpanda service. Test are
changed to use the default client instead. This is done either explicitly
(because the test is not using a high-level test harness) or they are and they
access it via `self.client()`.
