+++ tests/rptest/services/redpanda.py
@@ -701,28 +701,6 @@ class RedpandaService(Service):
                         counts[idx] += int(sample.value)
         return all(map(lambda count: count == 0, counts.values()))
 
-    def describe_topics(self, topics=None):
-        """
-        Describe topics. Pass topics=None to describe all topics, or a pass a
-        list of topic names to restrict the call to a set of specific topics.
-
-        Sample return value:
-            [
-              {'error_code': 0,
-               'topic': 'topic-kabn',
-               'is_internal': False,
-               'partitions': [
-                 {'error_code': 0,
-                  'partition': 0,
-                  'leader': 1,
-                  'replicas': [1],
-                  'isr': [1],
-                  'offline_replicas': []}
-               }
-            ]
-        """
-        return self._client.describe_topics(topics)
-
     def partitions(self, topic):
         """
         Return partition metadata for the topic.
